pyproject.toml
[project]
name = "iltci-presentation"
version = "0.1.0"
description = "Apply markdown content to ILTCI PowerPoint template"
requires-python = ">=3.8"
dependencies = [
    "python-pptx>=0.6.21",
    "pyyaml>=6.0",
    "streamlit>=1.38.0",
]

[project.scripts]
apply-template = "apply_content_to_template:main"